Refactor email handling through dedicated service
